moving all multithread logic to a mtcpclient package, that will be the one consuming the tcpclient, so main becomes clean. reporting is a functionality of this new package, isolated from the main thread runner
